Refactored out Dialog from UserDialog and CustomerDialog
